bot/requirements.txt
Scrapy==2.4.1
requests==2.25.1
treq==21.1.0
python-slugify==4.0.1